Intermediate Troubleshooting Commit
When compiling the document, the computer crashed. Since then the document does not compile anymore. This commit is to temporarily save the changes made, so that the document can be rolled back to a previous version. The goal is to see, if it does compile then, or if the library is now corrupted.