Bugfix and final Corrections Chap.02
Bugfix: Not sure what exactly the issue was. Maybe latex needed to compile other documents in between to reset for this repository. Maybe some other issue occured, that was fixed somehow through trial and error. However, now the bug seems to be fixed and the file compiles without errors (and the references seem to be linking correctly).

Corrections Chap.02: The last comments on mistakes in Chapter 02 are implemented. Only possibly missing is a literature review concerning error metrics for transient model calibration. This may be added later on.