Major correction Chap.04
Major corrections to Chapter 4 - Conceptual Model & Model Geometry. Almost all figures have been reviewed and replaced. Now, for each layer a map (if required) was generated and placed for presentation purposed wrapped in the text. For a satisfying placement in the text, much of the formatting had to be done manually.

Therefrom, a new ToDo stems now: The figure of the geological model was replaced last (Fig. 4.4). It needs to be resized (bigger) to be better readable. However doing this will influence the distribution of the paragraphs on the following pages. This will require a new manual placement and formatting of the wrapped figures. This may be done in a future commit.

So the following major ToDos are to be adressed in future in this chapter:
    - Resizing Figure 4.4, then re-placing of the following wrapped figures.
    - Adding more information in Section 4.6 Observations:
        - Full table of piezometers in appendix
        - Summary of information of that table in text
        - Possibly with an additional graph/map